Added activation key
User can now enable activation key to only get icon when they press the key
and text is highlighted.
Reintroduced wait period for seeing if user actually highlighted text.
This is necessary because sometimes the highlight is cancelled on a mouse
down event, other times on a mouse up event.